# Use `jax.numpy.diagonal` to evaluate `jax.numpy.trace`.