--- a/jax/_src/numpy/lax_numpy.py
+++ b/jax/_src/numpy/lax_numpy.py
@@ -3610,7 +3610,7 @@ def triu(m: ArrayLike, k: int = 0) -> Array:
 
 
 @util.implements(np.trace, skip_params=['out'])
-@partial(jit, static_argnames=('axis1', 'axis2', 'dtype'))
+@partial(jit, static_argnames=('offset', 'axis1', 'axis2', 'dtype'))
 def trace(a: ArrayLike, offset: int | ArrayLike = 0, axis1: int = 0, axis2: int = 1,
           dtype: DTypeLike | None = None, out: None = None) -> Array:
   util.check_arraylike("trace", a)
@@ -3618,13 +3618,14 @@ def trace(a: ArrayLike, offset: int | ArrayLike = 0, axis1: int = 0, axis2: int 
     raise NotImplementedError("The 'out' argument to jnp.trace is not supported.")
   dtypes.check_user_dtype_supported(dtype, "trace")
 
-  a_shape = shape(a)
-  a = moveaxis(a, (axis1, axis2), (-2, -1))
+  if dtype is None:
+    dtype = _dtype(a)
+    if issubdtype(dtype, integer):
+      default_int = dtypes.canonicalize_dtype(int)
+      if iinfo(dtype).bits < iinfo(default_int).bits:
+        dtype = default_int
 
-  # Mask out the diagonal and reduce.
-  a = where(eye(a_shape[axis1], a_shape[axis2], k=offset, dtype=bool),
-            a, zeros_like(a))
-  return reductions.sum(a, axis=(-2, -1), dtype=dtype)
+  return diagonal(a, offset=offset, axis1=axis1, axis2=axis2).sum(axis=-1, dtype=dtype)
 
 
 def _wrap_indices_function(f):