The current implementation of `jax.numpy.trace` uses masking to include/exclude the elements to trace. This can be slow, especially for larger arrays. This change instead uses `jax.numpy.diagonal` to extract the elements and evaluate the trace.

The performance gain depends on the size of the input arrays but can be >100x for larger arrays.

```python
# Use lambda to treat axis and offset as static args.
jitted_old = jax.jit(lambda x: old_trace(x))
jitted_new = jax.jit(lambda x: new_trace(x))

# Run the tests for arrays of different sizes.
key = jax.random.key(18)
for shape in [(10, 10), (100, 100, 8), (1000, 1000, 17)]:
    print(shape)
    key, subkey = jax.random.split(key)
    x = jax.random.normal(subkey, shape)
    for method, f in [("old", jitted_old), ("new", jitted_new)]:
        print(method)
        # Run once to ensure compiled for this shape.
        f(x).block_until_ready()
        # Run the timings.
        %timeit f(x).block_until_ready()
```
```
(10, 10)
old
4.17 µs ± 79.4 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)
new
3.41 µs ± 32.6 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)

(100, 100, 8)
old
107 µs ± 906 ns per loop (mean ± std. dev. of 7 runs, 10,000 loops each)
new
4.36 µs ± 61.8 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)

(1000, 1000, 17)
old
31.4 ms ± 1.21 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
new
35.4 µs ± 383 ns per loop (mean ± std. dev. of 7 runs, 10,000 loops each)
```