@jakevdp 2024-04-29
Thanks for the suggestion – it might be worth benchmarking on TPU as well. `gather` (which is the operation behind indexing) is typically very slow on TPU due to constraints of the hardware architecture, which may be why the masking implementation was chosen in the first place.