@tillahoffmann 2024-04-29
Interesting. Here are the results from a Colab TPU (with slightly larger trailing dimension to minimize profiling the overhead). It seems that the `diagonal`-based version still does quite well.

```
(10, 10, 500)
old
132 µs ± 2.37 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)
new
136 µs ± 2.02 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)
(100, 100, 500)
old
218 µs ± 7.32 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
new
132 µs ± 7.76 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)
(1000, 1000, 500)
old
7.07 ms ± 25.9 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
new
318 µs ± 19.1 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
```

Here's the full script I ran.

```python
import jax
from jax import numpy as jnp
from jax.numpy import trace as old_trace
import numpy as np


def new_trace(a, offset=0, axis1=0, axis2=1):
    return jnp.diagonal(a, offset=offset, axis1=axis1, axis2=axis2).sum(axis=-1)


# Use lambda to treat axis and offset as static args.
jitted_old = jax.jit(lambda x: old_trace(x))
jitted_new = jax.jit(lambda x: new_trace(x))

# Run the tests for arrays of different sizes.
key = jax.random.key(18)
for shape in [(10, 10, 500), (100, 100, 500), (1000, 1000, 500)]:
    print(shape)
    key, subkey = jax.random.split(key)
    x = jax.random.normal(subkey, shape)
    previous = None
    for method, f in [("old", jitted_old), ("new", jitted_new)]:
        print(method)
        # Run once to ensure compiled for this shape.
        y = f(x).block_until_ready()
        # Check the results are the same.
        if previous is not None:
          np.testing.assert_allclose(previous, y)
        # Run the timings.
        %timeit f(x).block_until_ready()
```