@hawkinsp 2024-04-29
Can you try dimensions 128x128 and 256x256 without a batch dimension, on colab TPU?

(I strongly suspect that on TPU within a block of size 128x128 the masking is optimal. If that were the case, then the optimal strategy is probably a hybrid of the two: gather diagonal blocks and then mask them. But it's possible XLA is smart enough to figure that out already.)