@tillahoffmann 2024-04-29
Looks like your hunch is right. I've included both with and without batch dimension. I suspect we may be profiling call overhead without the batch dimension because the time taken is almost constant.

**With batch dimension**

```
(32, 32, 500)
old
139 µs ± 3.18 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)
new
123 µs ± 6.9 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)
(64, 64, 500)
old
161 µs ± 5.02 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)
new
123 µs ± 5.4 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)
(128, 128, 500)
old
400 µs ± 65.5 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
new
940 µs ± 26.6 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
(256, 256, 500)
old
866 µs ± 13.6 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
new
1.92 ms ± 27.2 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
(512, 512, 500)
old
2.34 ms ± 36.3 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
new
7.77 ms ± 16.1 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
(1024, 1024, 500)
old
7.51 ms ± 26.5 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
new
29.1 ms ± 25 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
```

**Without batch dimension**

```
(32, 32)
old
126 µs ± 16.1 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)
new
116 µs ± 3.6 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)
(64, 64)
old
126 µs ± 8.95 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)
new
144 µs ± 16.8 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)
(128, 128)
old
112 µs ± 1.06 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)
new
136 µs ± 17.8 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)
(256, 256)
old
139 µs ± 18.4 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)
new
119 µs ± 4.76 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)
(512, 512)
old
132 µs ± 14.2 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)
new
133 µs ± 1.92 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)
(1024, 1024)
old
131 µs ± 5.61 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)
new
165 µs ± 2.61 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)
```